ParcialSI2AutomotrizBackend-main/ParcialSI2AutomotrizBackend-main/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
bcrypt==5.0.0
python-dotenv==1.2.2
python-multipart==0.0.24
pydantic==2.12.5
